rename components to page
