--- kitchen-ui/Order.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-function Order() {
-    return (
-        <div>
-            Order
-        </div>
-
-    )
-}
-
-export default Order;
